This changes the way OSDs are found and mounted
Before we were using filesystem uusid for naming and delegating the
whoami lookup to the osd.sh script. Now we are doing that lookup
outside of the container so that we can set the container name to the
OSD ID.

Additionally we are going bootstrapping sperately instead of on the
fly when the container started.