Removed port and volume declaration in Dockerfile
This information can quickly get out of date and requires a rebuild
to support a new port addition on a container like haproxy.

This also requires declaring the expose in the container starts due
to a possible bug in either ansible's docker, docker-py, or docker
api. Might not be a bug at all, might be desired behaviour.